[refactor] 파일 이름 오류와 동기 목록 초기화
