+++ src/components/StarRating.vue
